fix: 修复部分bug

+++ src/core/view.rs
@@ -237,7 +237,7 @@ impl View {
         let grapheme_delta = new_len.saturating_sub(old_len);
         if grapheme_delta > 0 {
             //move right for an added grapheme (should be the regular case)
-            self.move_right();
+            self.move_text_location(&Direction::Right);
         }
         self.need_redraw = true;
     }
@@ -283,7 +283,11 @@ impl View {
     }
 
     pub fn backspace(&mut self) {
-        self.move_left();
+        if self.text_location.line_index     == 0 && 
+           self.text_location.grapheme_index == 0 {
+            return ;
+        }
+        self.move_text_location(&Direction::Left);
         self.delete();
     }
